Backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
groq==0.5.0
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.0.0
pytest-asyncio>=0.23.5
